api/requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn
pydantic
